plugin: split basic blocks before and after a mv-call
In the assembly stage, GCC seems to collect all labels from a basic
block and places them before the assembly instruction of this basic
block. This leads to a soup of labels all pointing at the same address.
Therefore, we split the basic blocks before the inserted label and after
the call instruction and instruct gcc to not merge these blocks again.