callsites: find callsites in the RTL code
Callsites are much easier to detect and to equipt with a call-site label
in the RTL. Therefore, we move the detection of multiverse call sites
into an RTL pass. This simplifies the mv_info datastructures, since only
one label that points exactly to the callsite is emitted, and the libmv,
since we do not have to heuristically search for callsites.