gcc-plugin: merge mvfn descriptors, if possible
If multiple mvfn descriptors point to the same function body, we can
merge their respective descriptors, if their assignment maps are
compatible. By this we reduce the memory cost for descriptors and the
search for a fitting mvfn descriptor. The employed merge algorithm is
very simplistic and is currenlty only fitted for boolean assignment
maps.