plugin: variant generator and enums
The generation of variants for a multiverse function is now done via an
generator object, which also scored the single variants and is able to
limit the maximal number of generated variants (not implemented yet).
Furthermore, we also handle enumeral types, such that a variant for
every enum-elemnent is generated.