libmv: factor out arch specifica and detect trivial function bodies
First, we factor out the current patching and callsite-decoding
infrastructure into a architecture specific file. This will help future
portability.

Furthermore, we decode the multiverse function body. If it is
trivial (immediate return or return constant in eax), we special case
this and patch in a `mov constant, %eax' or a `nop' into the callsites.
This special treatment is especially useful for functionality, that is
fully disabled in one multiverse.