REAME: adapt readme to discussion and current implementation
The current implementation of multiverse does only support full
specialization of a function. This means that all referenced multiverse
variable are captured at commit time. We assume that all variables are
"stable". To reflect this current implementation state, I adapted the
README. After we have implemented partial specialization, we can revert
parts of the readme.