make: build EXTRA_DEPS first, split LDFLAGS from LDLIBS
When we're trying to execute a test-% target in a clean tree, we get
an error message because the plugin and library have not been built
before make tries to build the .o file for the testcase.
If we reorder the prerequisite list and shift $(EXTRA_DEPS) to the
front, we can make sure that extra dependencies are built first.

Additionally, the make documentation states that library names should
go into LDLIBS instead of LDFLAGS [0], so let's stick to that.

[0]: https://www.gnu.org/software/make/manual/html_node/Implicit-Variables.html