add docker build-test infrastructure
GCC is continuously changing, and compilation breaks every here and then
as GCC's API is everything but stable.  Many distributions don't ship
enough packages to build test the GCC plugin with different versions of
GCC as for specific versions dependencies are missing, code, the plugin
source, etc.  In other words, it's painful.

To ease our build process, add Docker support for testing multiverse.
The main Makefile is extended with several rules to build and run the
docker images, as specified in docker/Dockerfile.gcc*.

Currently there are two Dockerfiles (gcc6 & gcc7), both building ontop
of Ubuntu 16.10 and 17.04 respectively.  To run the images in a
container, simply type '$ make run-docker-gcc{6,7}' in the multiverse
root directory, which is then mounted as a volume in the container. This
allows us to build test the code as is, so that we can edit code on the
host and build test the changes in a container.  The images can also be
used for future automated tests.